Refactored code to separate encoding/decoding from backend logic
